Restructured project files, separating Labyrinth and respective JUnit files Added full coverage testing for LabyrinthMap.java in LabyrinthMapTest.java Made minor change for in-line return on Line 52 of LabyrinthMap.java
