--- src/BoardVisualTest.java
@@ -1,14 +1,0 @@
-package Labyrinth;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import org.junit.jupiter.api.Test;
-
-class BoardVisualTest {
-
-	@Test
-	void test() {
-		fail("Not yet implemented");
-	}
-
-}
