+++ src/Labyrinth/test/BoardVisualTest.java
@@ -1,0 +1,14 @@
+package Labyrinth.Labyrinth.test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class BoardVisualTest {
+
+	@Test
+	void test() {
+		fail("Not yet implemented");
+	}
+
+}
